mqtt_bridge/requirements.txt
paho-mqtt>=1.6.1,<1.7.0
influxdb-client>=1.37.0,<1.38.0
